removed temp page
fixed intro screen to work on the main app and removed the blank temp page between that
