+++ src/pages/Menu.tsx
@@ -1,13 +1,22 @@
 import { IonContent, IonHeader, IonItem, IonMenu, IonMenuToggle, IonPage, IonRouterOutlet, IonSplitPane, IonTitle, IonToolbar } from '@ionic/react';
-import React from 'react';
+import React, { useState } from 'react';
 import { Route, Redirect } from 'react-router';
 
 // page imports
 import CharacterList from './CharacterList';
 import Home from './Home';
+import Intro from '@/components/intro';
 
 
 const Menu: React.FC = () => {
+
+    // default false for intro seen
+    const [introSeen, setIntroSeen] = useState(false);
+
+    // sets seen intro to true after intro is finished
+    const finishIntro = async() => {
+        setIntroSeen(true);
+    }
 
     // list of pages here
     const paths = [
@@ -16,6 +25,10 @@ const Menu: React.FC = () => {
     ]
 
     return (
+    <>
+        {!introSeen ? (
+        <Intro onFinish={finishIntro} />
+        ) : (
         <IonPage>
             {/* split pane to hide navbar on smaller screen and have always open on larger screen */}
             <IonSplitPane contentId='main'> 
@@ -46,6 +59,8 @@ const Menu: React.FC = () => {
                 </IonRouterOutlet>
             </IonSplitPane>
         </IonPage>
+        )}
+    </>
     );
 };
